Reorganization of project, added tests
Fix #8
Fix #9

- Reorg of all code into single exportable /pkg folder
- Uncapitalized types not meant for export
- Added tests for SecretConfig and NewSecretClient. Further testing could
be achieved by hiding the tight http request integration behind an
interface that is provided to the HttpClient. In this way, requests can
be easily simulated. OR, if preferred, use the httptest package to
spin up a mock HTTP server for testing.

I think that with the above changes this module can be pulled into the
edgexfoundry organization. However it should probably be renamed to
something like go-mod-security-client.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>