Update key retrieval errors for multiple keys
Fix #6

Update the Go docs for 'GetValues' function to clearly express the
situations which will cause an error.

Renamed 'ErrSecretNotFound' to 'ErrSecretsNotFound' for consistency.

Update the 'ErrSecretsNotFound' error to contain multiple keys so that a
call to obtain multiple keys which fails can properly communicate which
keys has caused the error.

Add tests for the various errors which can be returned when obtaining
keys with multiple values.

Remove redundant tests.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>