Remove provider configuration property
Fix #18

Remove the SecretConfig's provider field which was previously used to
identify the SecretClient implementation which should be use during
creation. This is not necessary as there is currently only one provider
for Vault and this generalization is most likely a concern for consumers
of the go-mod-secrets module as this still would not solve the issue of
abstracting the creation of SecretClients since there could be clients
which live outside of this module.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>