Update SecretClient provider ID to be explicit
Update the SecretClient constructor to use a more explicit identifier
for the Vault provider. Previously, the provider used was 'http'. From
the callers perspective this is not completely deterministic as to what
secret store abstraction is being used. It is more clear to expect the
caller to provide a 'provider' which indicates which abstraction they
would like to use and have the 'protocol' indicate the transport
mechanism which the provider should use. A good example  of this would
be if we support different versions of the Vault API, we could have
different providers which expect different response structures but the
communication protocol could be the same and the user could use either
depending on their situation.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>