Add Vault Namespace support within SecretClient
Fix #17

Update the SecretConfig and Vault SecretClient to support Vault's notion
of Namespaces. The SecretConfig has been updated to contain a
configuration property which allow users to specify the Namespace to
operate under when interacting with the Vault SecretStore. The
SecretClient has been updated to provide the Namespace information in
the appropriate header when interacting with the Vault HTTP API

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>