Update SecretClient to use path argument
Fix #26

Update the SecretClient and SecretStoreManager to accept a 'path'
argument which can be used to further specify the type/location of
secrets. This will allow the user to use one instance of the
SecretClient for obtaining different types of secrets. Previously this
value was only specified in the SecretConfig which would require
multiple instances if the value needed to be changed in order to obtain
other secrets.

Update the Vault implementation to accept 0 keys which would result in
all of the secrets being passed back to the caller. Previously, this
would result in the caller getting no secrets. This is now more of a
filter to allow the caller to limit the amount of keys being returned.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>