Refactor SecretClient from struct to interface
Simplify the SecretClient by making it an interface rather than a struct
to eliminate the unnecessary layer of indirection. The SecretClient was
first designed to contain more methods which would wrap a
SecretStoreManager interface allowing future additions to only have
implement a smaller sub-set of methods needed by the SecretStoreManager
interface. Since then we have removed all of the methods in the
SecretClient and SecretStoreManager except for one. These changes go
against the original intent of the SecretClient and SecretStoreManager
and now are not needed. These changes do not affect any external APIs
and will not affect consumers who create SecretClients via a constructor
and use the SecretClient interface.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>