Update SecretClient to use path argument and Refactor SecretClient from struct to interface (#31)
* Update SecretClient to use path argument

Fix #26

Update the SecretClient and SecretStoreManager to accept a 'path'
argument which can be used to further specify the type/location of
secrets. This will allow the user to use one instance of the
SecretClient for obtaining different types of secrets. Previously this
value was only specified in the SecretConfig which would require
multiple instances if the value needed to be changed in order to obtain
other secrets.

Update the Vault implementation to accept 0 keys which would result in
all of the secrets being passed back to the caller. Previously, this
would result in the caller getting no secrets. This is now more of a
filter to allow the caller to limit the amount of keys being returned.

* Refactor SecretClient from struct to interface

Simplify the SecretClient by making it an interface rather than a struct
to eliminate the unnecessary layer of indirection. The SecretClient was
first designed to contain more methods which would wrap a
SecretStoreManager interface allowing future additions to only have
implement a smaller sub-set of methods needed by the SecretStoreManager
interface. Since then we have removed all of the methods in the
SecretClient and SecretStoreManager except for one. These changes go
against the original intent of the SecretClient and SecretStoreManager
and now are not needed. These changes do not affect any external APIs
and will not affect consumers who create SecretClients via a constructor
and use the SecretClient interface.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>