pkg/provides/vault/many: allow specifying retry logic
Allow a SecretClient to be created that retries HTTP requests that are
not 2XX. Note that this behavior is optional, and needs to be opt-in by
the calling client.

This enables clients of go-mod-secrets to easily block waiting for their
secrets to be available.

Also include a proposed implementation of "infinite" blocking by setting
the AdditionalRetryAttempts to a negative number, but error out for now,
so as to not complicate testing.

Finally, update the tests to check the number of calls to Do() that are
made when getting secrets under various situations.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>