vault/client.go: refactor GetSecrets to do the retry loop instead of getAllKeys
This means that we will also retry catastrophic errors like not being able to
allocate memory to write to the socket, etc. in addition to non-200 HTTP status
codes returned from Vault.

This also means we need to change the tests a bit, now we will retry Do() calls
even if all errors are "catastrophic" and not HTTP status codes.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>