pkg/providers/vault: change RetryWaitPeriod to string, do parsing here
It is simpler for clients of go-mod-secrets to just provide a string here since
in many cases the SecretConfig struct will be built verbatim from a
configuration.toml file, where we don't have complex types like time.Duration
(or at least a usable version of that - time.Duration serializes as an int64 in
nanoseconds which is very unusable to specify in the configuration.toml).

This means we need to do the parsing of the string here if it exists and error
in the case parsing fails.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>